requirements.txt
streamlit>=1.27.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.3.0
pypdf>=3.15.2
python-docx>=0.8.11  # For Word document support
tiktoken>=0.5.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
chromadb>=0.4.13
sentence-transformers>=2.2.2
nltk>=3.8.1
spacy>=3.7.2
deep-translator>=1.11.4
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
fpdf2>=2.7.5  # For PDF export
wordcloud>=1.9.2
seaborn>=0.12.0
matplotlib>=3.7.0
pandas>=1.5.0
textblob>=0.17.1
nltk>=3.8.1